# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Just Her Rideshare",
        "description": "Just Her Rideshare",
        "contact": {
            "name": "Smart Data Team"
        }
    },
    "paths": {
...
        "/auth/check-email-mobile-exists": {
            "post": {
                "x-controller-name": "AuthController",
                "x-operation-name": "CheckEMailMobileExist",
                "tags": ["AuthController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "CheckEMailMobileExist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "mobile_country_code": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "user_type": {
                                        "type": "string",
                                        "enum": ["customer", "driver"]
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "AuthController.CheckEMailMobileExist"
            }
        },
        "/auth/sign-in": {
            "post": {
                "x-controller-name": "AuthController",
                "x-operation-name": "signIn",
                "tags": ["AuthController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "signIn",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "mobile_country_code": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "user_type": {
                                        "type": "string",
                                        "enum": ["customer", "driver"]
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "AuthController.signIn"
            }
        },
        "/auth/sign-up": {
            "post": {
                "x-controller-name": "AuthController",
                "x-operation-name": "SignUp",
                "tags": ["AuthController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "SignUp",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "mobile_country_code": {
                                        "type": "string"
                                    },
                                    "profile_picture": {
                                        "type": "file",
                                        "description": "As a multipart"
                                    },
                                    "user_type": {
                                        "type": "string",
                                        "enum": ["customer", "driver"]
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": ["male", "female"]
                                    },
                                    "state_id": {
                                        "type": "string"
                                    },
                                    "city_id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "AuthController.SignUp"
            }
        },
        "/auth/verify-otp": {
            "post": {
                "x-controller-name": "AuthController",
                "x-operation-name": "VerifyOTPLogin",
                "tags": ["AuthController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "VerifyOTPLogin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "user_type": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "otp": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "AuthController.VerifyOTPLogin"
            }
        },
        "/auth/resend-otp": {
            "post": {
                "x-controller-name": "AuthController",
                "x-operation-name": "ResendOtpLogin",
                "tags": ["AuthController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "ResendOtpLogin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "user_type": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "AuthController.ResendOtpLogin"
            }
        },
        "/auth/verify-mobile": {
            "post": {
                "x-controller-name": "AuthController",
                "x-operation-name": "VerifyMobile",
                "tags": ["AuthController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "VerifyMobile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "verify_mobile_id": {
                                        "type": "string"
                                    },
                                    "otp": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "AuthController.VerifyMobile"
            }
        },
        "/auth/resend-otp-verify-mobile": {
            "post": {
                "x-controller-name": "AuthController",
                "x-operation-name": "ResendOtpVerifyMobile",
                "tags": ["AuthController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "ResendOtpVerifyMobile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "verify_mobile_id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "AuthController.ResendOtpVerifyMobile"
            }
        },
        "/auth/logout-delete-account": {
            "get": {
                "x-controller-name": "AuthController",
                "x-operation-name": "LogoutDelete",
                "tags": ["AuthController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "LogoutDelete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "code",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "user_type",
                    "in": "query",
                    "schema": {
                        "type": "string",
                        "enum": ["customer", "driver"]
                    }
                }, {
                    "name": "action",
                    "in": "query",
                    "schema": {
                        "type": "string",
                        "enum": ["logout", "delete"]
                    }
                }, {
                    "name": "user_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "AuthController.LogoutDelete"
            }
        },
        "/auth/get-profile": {
            "get": {
                "x-controller-name": "AuthController",
                "x-operation-name": "GetProfile",
                "tags": ["AuthController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "GetProfile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "code",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "user_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "user_type",
                    "in": "query",
                    "schema": {
                        "type": "string",
                        "enum": ["customer", "driver"]
                    }
                }],
                "operationId": "AuthController.GetProfile"
            }
        },
        "/auth/update-profile": {
            "post": {
                "x-controller-name": "AuthController",
                "x-operation-name": "UpdateProfile",
                "tags": ["AuthController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "UpdateProfile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "user_type": {
                                        "type": "string",
                                        "enum": ["customer", "driver"]
                                    },
                                    "profile_picture": {
                                        "type": "file",
                                        "description": "file as a multipart"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "AuthController.UpdateProfile"
            }
        },
...
    },
...
    "servers": [{
        "url": "http://52.70.114.84:9133/api"
    }, {
        "url": "http://localhost:9133/api"
    }]
}